requirements.txt
setuptools
dataclass_csv
google-cloud-spanner
numpy
alive_progress
Faker
